// src/api/generated/schemas/device.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * NervesCloud API
 * The NervesCloud API gives users full access to their
Orgs, Products, and corresponding Device fleets.

The API can be used to integrate with your own systems, providing full access to your Product and Device data.

The API is documented using the OpenAPI 3.0 specification.

 * OpenAPI spec version: 2.0.0
 */
import type { DeviceConnectionStatus } from './deviceConnectionStatus';
import type { DeviceDeploymentGroup } from './deviceDeploymentGroup';
import type { DeviceFirmwareMetadata } from './deviceFirmwareMetadata';

export interface Device {
  connection_status?: DeviceConnectionStatus;
  deployment_group?: DeviceDeploymentGroup;
  description?: string;
  firmware_metadata?: DeviceFirmwareMetadata;
  identifier?: string;
  /** @deprecated */
  last_communication?: string;
  online?: boolean;
  org_name?: string;
  product_name?: string;
  tags?: string[];
  /** Device penalty box expiration timestamp */
  updates_blocked_until?: string;
  updates_enabled?: boolean;
  version?: string;
  /** @nullable */
  id?: number | null;
  deleted?: boolean;
}