// src/api/generated/schemas/deviceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NervesHub API
 * OpenAPI spec version: 2.0.0
 */
import type { Device } from './device';
import type { Pagination } from './pagination';

/**
 * Response schema for multiple Devices
 */
export interface DeviceListResponse {
  /** The Device schema */
  data?: Device[];
  pagination?: Pagination;
}